code: 문자열을 잘라서 합하는 기능이 있는 메서드 추가
