--- src/main/java/calculator/calculator.java
@@ -1,16 +1,0 @@
-package calculator;
-
-public class calculator {
-    int add(int num1, int num2){
-        return num1+num2;
-    }
-    int subtract(int num1, int num2){
-        return num1-num2;
-    }
-    int multiply(int num1, int num2){
-        return num1*num2;
-    }
-    int divide(int num1, int num2){
-        return num1/num2;
-    }
-}
